fix: Correctly handle indentation for nested json
Previously the recursive call to sStringify would take
already stringified nested object and append only one indent
to account for it being a child of another one.

This is incorrect

If we are given a nested object string and are placing it
as a child of another object, we need to prepend the indent
string to each of the lines of the nested object string
apart from the open and end tokens (e.g. so anything matching
\{}[]\)

This is a naive fix